Separate some common code related to TWKB.
